feat: convert mongooser into functional components
