+++ src/models/base-document.ts
@@ -1,5 +1,8 @@
+// libs
 import { Typegoose } from 'typegoose';
 
+export type UniqueId = any;
+
 export class BaseDocument extends Typegoose {
-  _id: any;
+  _id: UniqueId;
 }
